Move all constants to own folder
