+++ server/constants/client.ts
@@ -1,0 +1,3 @@
+export const CLIENTROLES = {
+  client: "client",
+} as const;
