+++ server/constants/employee.ts
@@ -1,0 +1,4 @@
+export const ROLES = {
+  admin: "admin",
+  support_agent: "support-agent",
+} as const;
