+++ server/constants/refunds.ts
@@ -1,0 +1,6 @@
+export const REFUNDSTATUS = {
+  pending: "pending",
+  processing: "processing",
+  accepted: "accepted",
+  declined: "declined",
+} as const;
